feat: add CompanyBox component for displaying company details and booking sessions
